changed cash calculations to new formula
this commit also changed twnetyfivek protection to be a variable amount
of protection (right now at 1500), and changed my overall limit to 500.
I've realized that as long as the ACH actually subtracts the 1500 from my cash
in the morning, (assuming I will have made that much), I can during the day cancel the program
restart with 1500 protection off, and continue. So no need to stop (hopefully),
thus tagged unstoppable. (this prolly should've gone in the tag annotation but oh well)