fixed illegal instruction error in ChangeUpTheFiles func.
problem was the func. was of type int. but didn't have a return statement anywhere. May go back to holding overnight with a cash reserve after running this for a couple weeks or so if i find i'm loosing too much profit from the overnight drop perhaps somewhat commonly seen?